Allow CLI to communicate with production app
By default, the stethoscope CLI will pass along an origin value of
'stethoscope://main' to validate(). This is the default allowedHosts
configuration, and if unchanged, will allow the CLI to talk to
production builds.

If a build uses custom allowedHosts values, this header should be
configurable, so the CLI accepts an --origin flag.

Also, ensure 'node-fetch' is available.