Firewall kmd source for linux (#106)
Here we're assuming that the desktop user has the `ufw` package
installed, and UFW is enabled. This seems like a sane default choice
though there are many ways to do firewalling on linux. Toggling UFW
status will reliably toggle 0<->1 for the ipt_REJECT and ip6t_REJECT
kernel module states for ipv4 and ipv6. If the UFW package is not
installed, this will almost certainly return `undefined`.