Implement applications resolver for win32 (#170)
In this resolver context, specifying a policy like:

{
  'applications': [
    {
      'name': 'someapp',
      'assertion': 'ALWAYS',
      'paths': {
        'win32': '\\\\SOME\\\\REGISTRY\\\\PATH'
      }
    }
  ]
}

the 'paths.win32' string here indicates that the resolver should look in
that registry location for installation data. Most likely, this
shouldn't need to be specified, and the querent may rely on the default
resolver registry path.

NOTE: platform filtering is not yet working. The applications policy
schema needs an update which will occur under a separate PR.